(maint) Move host processing logic to the presuites that need updating

+++ setup/common/040_ValidateSignCert.rb
@@ -1,19 +1,23 @@
 test_name "Validate Sign Cert" do
   skip_test 'not testing with puppetserver' unless @options['is_puppetserver']
-  hostname = on(master, 'facter hostname').stdout.strip
-  fqdn = on(master, 'facter fqdn').stdout.strip
-  puppet_version = on(master, puppet("--version")).stdout.chomp
+  hostname = on(master, 'facter hostname').stdout.strip unless master['use_existing_container'] == 'true'
+  fqdn = on(master, 'facter fqdn').stdout.strip unless master['use_existing_container'] == 'true'
+  puppet_version = on(master, puppet("--version")).stdout.chomp unless master['use_existing_container'] == 'true'
 
   if master.use_service_scripts?
     step "Ensure puppet is stopped"
     # Passenger, in particular, must be shutdown for the cert setup steps to work,
     # but any running puppet master will interfere with webrick starting up and
     # potentially ignore the puppet.conf changes.
-    on(master, puppet('resource', 'service', master['puppetservice'], "ensure=stopped"))
+    on(master, puppet('resource', 'service', master['puppetservice'], "ensure=stopped")) unless master['use_existing_container'] == 'true'
   end
 
   step "Clear SSL on all hosts"
   hosts.each do |host|
+    if host['use_existing_container'] == 'true'
+      puts "SKIPPING CLEARING SSL!"
+      next
+    end
     ssldir = on(host, puppet('agent --configprint ssldir')).stdout.chomp
     # preserve permissions for master's ssldir so puppetserver can read it
     on(host, "rm -rf #{ssldir}/*")
@@ -35,7 +39,7 @@ test_name "Validate Sign Cert" do
 
     # In Puppet 6, we want to be using an intermediate CA
     unless version_is_less(puppet_version, "5.99")
-      on master, 'puppetserver ca setup'
+      on master, 'puppetserver ca setup' unless master['use_existing_container'] == 'true'
     end
     with_puppet_running_on(master, master_opts) do
       step "Agents: Run agent --test with autosigning enabled to get cert"
